refactor:ReservationService를 repository로 변경
