+++ src/main/java/roomescape/controller/ReservationController.java
@@ -9,7 +9,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-import roomescape.service.ReservationService;
+import roomescape.persistence.ReservationRepository;
 
 import java.util.List;
 import java.util.Map;
@@ -18,15 +18,15 @@ import java.util.Map;
 @RequestMapping("/reservations")
 public class ReservationController {
 
-    private final ReservationService reservationService;
+    private final ReservationRepository reservationRepository;
 
-    public ReservationController(ReservationService reservationService) {
-        this.reservationService = reservationService;
+    public ReservationController(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
     }
 
     @GetMapping
     public ResponseEntity<List<ReservationDto>> getAllReservations() {
-        List<ReservationDto> reservations = reservationService.findAll();
+        List<ReservationDto> reservations = this.reservationRepository.findAll();
 
         return ResponseEntity.ok()
                 .body(reservations);
@@ -34,7 +34,7 @@ public class ReservationController {
 
     @PostMapping
     public ResponseEntity<ReservationDto> createReservations(@RequestBody Map<String, String> reservationRequest) {
-        ReservationDto reservation = reservationService.save(reservationRequest);
+        ReservationDto reservation = reservationRepository.save(reservationRequest);
 
         return ResponseEntity.status(HttpStatus.CREATED)
                 .header("Location", "/reservations/" + reservation.id())
@@ -43,7 +43,7 @@ public class ReservationController {
 
     @DeleteMapping("/{deleteId}")
     public ResponseEntity<Void> deleteById(@PathVariable long deleteId) {
-        reservationService.deleteById(deleteId);
+        reservationRepository.deleteById(deleteId);
 
         return ResponseEntity.status(HttpStatus.NO_CONTENT)
                 .build();
